requirements.txt
anyio==3.7.1
APScheduler==3.6.3
black==23.7.0
cachetools==4.2.2
certifi==2023.7.22
click==8.1.6
colorama==0.4.6
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
requests==2.26.0